+++ src/Alexa/Response/Directives/AbstractDirective.php
@@ -1,0 +1,31 @@
+<?php
+
+namespace InternetOfVoice\LibVoice\Alexa\Response\Directives;
+
+/**
+ * Abstract Class AbstractDirective
+ *
+ * @author  Alexander Schmidt <a.schmidt@internet-of-voice.de>
+ * @license http://opensource.org/licenses/MIT
+ */
+abstract class AbstractDirective {
+	/** @var string $type */
+	protected $type;
+
+
+	public function __construct() {
+	}
+
+
+	/**
+	 * @return string
+	 */
+	public function getType() {
+		return $this->type;
+	}
+
+	/**
+	 * @return array
+	 */
+	abstract function render();
+}