+++ CHANGELOG.md
@@ -2,6 +2,13 @@
 
 Changes are documented in reverse chronological order.
 
+## Version 1.2.0 (2017-12-19)
+* Add AudioPlayer.* requests
+* Add PlaybackController.* requests
+* Add System.ExceptionEncountered request
+* Add VideoApp directive
+* Extend Alexa Smart Home Skill API (Beta)
+
 ## Version 1.1.0 (2017-11-29)
 * Add Alexa Smart Home Skill API (Beta)